🎨️ Rename : "Auth 도메인 패키지 및 코드 구조 조정"
- 외부 API 사용하는 검증 인가는 인프라 쪽으로 조정하였습니다 :)

Ref : #94
