+++ src/main/java/alcoholboot/toastit/infra/email/service/EmailService.java
@@ -1,4 +1,4 @@
-package alcoholboot.toastit.auth.email.service;
+package alcoholboot.toastit.infra.email.service;
 
 public interface EmailService {
     void sendSimpleMessage(String to, String subject, String text);
